requirements.txt
django~=1.8.0
django-extensions==1.5.7
djangorestframework==3.2.3
dj-database-url==0.4.1
gunicorn==19.3.0
python-dotenv==0.1.3
pyyaml==3.11
hashids==1.1.0
psycopg2==2.6.1
-i https://pypi-3844.21.co/pypi 
two1
